backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
python-dotenv==1.0.0
llama-index==0.10.18
llama-index-llms-openllm
openai
Flask==2.2.3
Flask-Cors==3.0.10
python-dotenv==1.0.0
llama-index==0.10.18
llama-index-llms-groq==0.1.3
groq==0.4.2
llama-index-embeddings-huggingface==0.2.0
transformers
torch  # required for transformers/HuggingFace models
